+++ src/components/[guild]/EditGuild/components/Admins/Admins.tsx
@@ -5,17 +5,31 @@ import {
   FormLabel,
   Tag,
 } from "@chakra-ui/react"
+import { HStack, Icon, forwardRef } from "@chakra-ui/react"
+import { Warning } from "@phosphor-icons/react"
+import {
+  CreatableSelect,
+  GroupBase,
+  MultiValueGenericProps,
+  Props,
+  chakraComponents,
+} from "chakra-react-select"
 import useGuild from "components/[guild]/hooks/useGuild"
 import useGuildPermission from "components/[guild]/hooks/useGuildPermission"
 import useUser from "components/[guild]/hooks/useUser"
+import CopyableAddress from "components/common/CopyableAddress"
 import GuildAvatar from "components/common/GuildAvatar"
+import StyledSelect from "components/common/StyledSelect"
+import CustomMenuList from "components/common/StyledSelect/components/CustomMenuList"
 import { useMemo } from "react"
-import { useController, useFormContext } from "react-hook-form"
+import { PropsWithChildren, useEffect } from "react"
+import { useController, useFormContext, useWatch } from "react-hook-form"
 import { FUEL_ADDRESS_REGEX, SelectOption } from "types"
 import { ADDRESS_REGEX } from "utils/guildCheckout/constants"
 import shortenHex from "utils/shortenHex"
+import { useEnsAddress } from "wagmi"
+import { mainnet } from "wagmi/chains"
 import TransferOwnership from "../TransferOwnership"
-import AdminSelect from "./components/AdminSelect"
 
 export const isValidAddress = (address: string) =>
   ADDRESS_REGEX.test(address) || FUEL_ADDRESS_REGEX.test(address)
@@ -126,5 +140,85 @@ const Admins = () => {
     </>
   )
 }
+
+type PropsHelper = MultiValueGenericProps<unknown, boolean, GroupBase<unknown>>
+const CustomMultiValueLabel = (props: PropsWithChildren<PropsHelper>) => {
+  const domain = isValidAddress(props.data.value) ? undefined : props.data.value
+  const { data: resolvedAddress } = useEnsAddress({
+    name: domain,
+    chainId: mainnet.id,
+  })
+
+  const { setError, setValue, control, trigger } = useFormContext()
+  const admins = useWatch({ control: control, name: "admins" })
+
+  useEffect(() => {
+    if (
+      domain &&
+      resolvedAddress &&
+      !admins.includes(resolvedAddress.toLowerCase())
+    ) {
+      setValue(
+        "admins",
+        admins.map((admin) =>
+          admin.address === domain ? { address: resolvedAddress } : admin
+        )
+      )
+
+      trigger("admins")
+    }
+
+    if (resolvedAddress === null) {
+      setError("admins", {
+        message: "Resolving address failed",
+      })
+    }
+  }, [domain, resolvedAddress, admins, setValue, trigger, setError])
+
+  return (
+    <chakraComponents.MultiValueLabel {...props}>
+      <HStack gap="0">
+        {resolvedAddress === null ||
+        admins.includes(resolvedAddress?.toLowerCase()) ? (
+          <Icon as={Warning} mr={1} color="red.300" boxSize={4} weight="bold" />
+        ) : (
+          props.data.img
+        )}
+        <CopyableAddress
+          address={props.data.value}
+          domain={domain ?? resolvedAddress}
+          size="sm"
+        />
+      </HStack>
+    </chakraComponents.MultiValueLabel>
+  )
+}
+
+const AdminSelect = forwardRef((props: Props, ref) => (
+  <StyledSelect
+    as={CreatableSelect}
+    size="lg"
+    components={{
+      MultiValueLabel: CustomMultiValueLabel,
+      Input: (inputProps) => (
+        <chakraComponents.Input {...inputProps} placeholder="Paste address" />
+      ),
+      DropdownIndicator: () => null,
+      MenuList: (props) => <CustomMenuList {...props} noResultText="No members" />,
+    }}
+    openMenuOnClick={false}
+    ref={ref}
+    {...props}
+    chakraStyles={{
+      valueContainer: (base) => ({ ...base, py: 2, px: 3 }),
+      multiValue: (base) => ({ ...base, minH: "7" }),
+      input: (provided) => ({
+        ...provided,
+        pl: 1,
+        minWidth: "14ch",
+      }),
+    }}
+  />
+))
 
 export default Admins
